+++ CHANGELOG.md
@@ -22,6 +22,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * The new `jj print` command prints the contents of a file in a revision.
 
+* (#243) The new `jj git remotes list` command lists the configured remotes and
+  their URLs.
+
 * `jj move` and `jj squash` now lets you limit the set of changes to move by
   specifying paths on the command line (in addition to the `--interactive`
   mode). For example, use `jj move --to @-- foo` to move the changes to file